ISSUE #4777 move schema to schema folder
